move note record structs to its own module
