+++ src/store/mod.rs
@@ -1,6 +1,6 @@
 use crate::{
     client::transactions::{TransactionRecord, TransactionResult},
-    errors::{ClientError, StoreError},
+    errors::StoreError,
 };
 use clap::error::Result;
 use miden_objects::{
@@ -9,20 +9,19 @@ use miden_objects::{
         dsa::rpo_falcon512::KeyPair,
         merkle::{InOrderIndex, MmrPeaks},
     },
-    notes::{
-        Note, NoteAssets, NoteId, NoteInclusionProof, NoteInputs, NoteMetadata, NoteScript,
-        Nullifier,
-    },
-    transaction::{InputNote, TransactionId},
+    notes::{NoteId, NoteInclusionProof, Nullifier},
+    transaction::TransactionId,
     utils::collections::BTreeMap,
-    BlockHeader, Digest, NoteError, Word,
+    BlockHeader, Digest, Word,
 };
-use serde::{Deserialize, Serialize};
 
 use miden_tx::utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
 pub mod data_store;
 pub mod sqlite_store;
+
+mod note_record;
+pub use note_record::{InputNoteRecord, NoteRecordDetails, NoteStatus, OutputNoteRecord};
 
 #[cfg(any(test, feature = "mock"))]
 pub mod mock_executor_data_store;
@@ -271,389 +270,6 @@ impl Deserializable for AuthInfo {
             }
             val => Err(DeserializationError::InvalidValue(val.to_string())),
         }
-    }
-}
-
-// NOTE STATUS
-// ================================================================================================
-#[derive(Clone, Copy, Debug, Serialize, Deserialize, PartialEq, Eq)]
-pub enum NoteStatus {
-    Pending,
-    Committed,
-    Consumed,
-}
-
-impl From<NoteStatus> for u8 {
-    fn from(value: NoteStatus) -> Self {
-        match value {
-            NoteStatus::Pending => 0,
-            NoteStatus::Committed => 1,
-            NoteStatus::Consumed => 2,
-        }
-    }
-}
-
-impl TryFrom<u8> for NoteStatus {
-    type Error = DeserializationError;
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            0 => Ok(NoteStatus::Pending),
-            1 => Ok(NoteStatus::Committed),
-            2 => Ok(NoteStatus::Consumed),
-            _ => Err(DeserializationError::InvalidValue(value.to_string())),
-        }
-    }
-}
-
-impl Serializable for NoteStatus {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        target.write_bytes(&[(*self).into()]);
-    }
-}
-
-impl Deserializable for NoteStatus {
-    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
-        let enum_byte = u8::read_from(source)?;
-        enum_byte.try_into()
-    }
-}
-
-// INPUT NOTE RECORD
-// ================================================================================================
-
-/// Represents a Note of which the [Store] can keep track and retrieve.
-///
-/// An [InputNoteRecord] contains all the information of a [Note], in addition of (optionally)
-/// the [NoteInclusionProof] that identifies when the note was included in the chain. Once the
-/// proof is set, the [InputNoteRecord] can be transformed into an [InputNote] and used as input
-/// for transactions.
-#[derive(Clone, Debug, PartialEq)]
-pub struct InputNoteRecord {
-    assets: NoteAssets,
-    details: NoteRecordDetails,
-    id: NoteId,
-    inclusion_proof: Option<NoteInclusionProof>,
-    metadata: Option<NoteMetadata>,
-    recipient: Digest,
-    status: NoteStatus,
-}
-
-impl InputNoteRecord {
-    pub fn new(
-        id: NoteId,
-        recipient: Digest,
-        assets: NoteAssets,
-        status: NoteStatus,
-        metadata: Option<NoteMetadata>,
-        inclusion_proof: Option<NoteInclusionProof>,
-        details: NoteRecordDetails,
-    ) -> InputNoteRecord {
-        InputNoteRecord {
-            id,
-            recipient,
-            assets,
-            status,
-            metadata,
-            inclusion_proof,
-            details,
-        }
-    }
-
-    pub fn id(&self) -> NoteId {
-        self.id
-    }
-
-    pub fn recipient(&self) -> Digest {
-        self.recipient
-    }
-
-    pub fn assets(&self) -> &NoteAssets {
-        &self.assets
-    }
-
-    pub fn status(&self) -> NoteStatus {
-        self.status
-    }
-
-    pub fn metadata(&self) -> Option<&NoteMetadata> {
-        self.metadata.as_ref()
-    }
-
-    pub fn nullifier(&self) -> &str {
-        &self.details.nullifier
-    }
-
-    pub fn inclusion_proof(&self) -> Option<&NoteInclusionProof> {
-        self.inclusion_proof.as_ref()
-    }
-
-    pub fn details(&self) -> &NoteRecordDetails {
-        &self.details
-    }
-}
-
-impl Serializable for InputNoteRecord {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        self.id().write_into(target);
-        self.recipient().write_into(target);
-        self.assets().write_into(target);
-        self.status().write_into(target);
-        self.metadata().write_into(target);
-        self.details().write_into(target);
-        self.inclusion_proof().write_into(target);
-    }
-}
-
-impl Deserializable for InputNoteRecord {
-    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
-        let id = NoteId::read_from(source)?;
-        let recipient = Digest::read_from(source)?;
-        let assets = NoteAssets::read_from(source)?;
-        let status = NoteStatus::read_from(source)?;
-        let metadata = Option::<NoteMetadata>::read_from(source)?;
-        let details = NoteRecordDetails::read_from(source)?;
-        let inclusion_proof = Option::<NoteInclusionProof>::read_from(source)?;
-
-        Ok(InputNoteRecord {
-            id,
-            recipient,
-            assets,
-            status,
-            metadata,
-            inclusion_proof,
-            details,
-        })
-    }
-}
-
-impl From<Note> for InputNoteRecord {
-    fn from(note: Note) -> Self {
-        InputNoteRecord {
-            id: note.id(),
-            recipient: note.recipient(),
-            assets: note.assets().clone(),
-            status: NoteStatus::Pending,
-            metadata: Some(*note.metadata()),
-            inclusion_proof: None,
-            details: NoteRecordDetails {
-                nullifier: note.nullifier().to_string(),
-                script: note.script().to_bytes(),
-                inputs: note.inputs().to_bytes(),
-                serial_num: note.serial_num(),
-            },
-        }
-    }
-}
-
-impl From<InputNote> for InputNoteRecord {
-    fn from(recorded_note: InputNote) -> Self {
-        InputNoteRecord {
-            id: recorded_note.note().id(),
-            recipient: recorded_note.note().recipient(),
-            assets: recorded_note.note().assets().clone(),
-            status: NoteStatus::Pending,
-            metadata: Some(*recorded_note.note().metadata()),
-            details: NoteRecordDetails {
-                nullifier: recorded_note.note().nullifier().to_string(),
-                script: recorded_note.note().script().to_bytes(),
-                inputs: recorded_note.note().inputs().to_bytes(),
-                serial_num: recorded_note.note().serial_num(),
-            },
-            inclusion_proof: Some(recorded_note.proof().clone()),
-        }
-    }
-}
-
-impl TryInto<InputNote> for InputNoteRecord {
-    type Error = ClientError;
-
-    fn try_into(self) -> Result<InputNote, Self::Error> {
-        match (self.inclusion_proof, self.metadata) {
-            (Some(proof), Some(metadata)) => {
-                let script = NoteScript::read_from_bytes(&self.details.script).map_err(|err| {
-                    ClientError::NoteError(NoteError::NoteDeserializationError(err))
-                })?;
-                let inputs = NoteInputs::read_from_bytes(&self.details.inputs).map_err(|err| {
-                    ClientError::NoteError(NoteError::NoteDeserializationError(err))
-                })?;
-                let note = Note::from_parts(
-                    script,
-                    inputs,
-                    self.assets,
-                    self.details.serial_num,
-                    metadata,
-                );
-                Ok(InputNote::new(note, proof.clone()))
-            }
-            // TODO: should we use a better Error for these two?
-            (None, _) => Err(ClientError::NoteError(
-                miden_objects::NoteError::invalid_origin_index(
-                    "Input Note Record contains no inclusion proof".to_string(),
-                ),
-            )),
-            (_, None) => Err(ClientError::NoteError(
-                miden_objects::NoteError::invalid_origin_index(
-                    "Input Note Record contains no metadata".to_string(),
-                ),
-            )),
-        }
-    }
-}
-
-// OUTPUT NOTE RECORD
-// ================================================================================================
-
-/// Represents a Note which was the result of executing some transaction of which the [Store] can
-/// keep track and retrieve.
-///
-/// An [OutputNoteRecord] contains all the information of a [Note] while it allows for not
-/// knowing the details (nullifier, script, inputs and serial number), in addition of (optionally)
-/// the [NoteInclusionProof] that identifies when the note was included in the chain.
-#[derive(Clone, Debug, PartialEq)]
-pub struct OutputNoteRecord {
-    assets: NoteAssets,
-    details: Option<NoteRecordDetails>,
-    id: NoteId,
-    inclusion_proof: Option<NoteInclusionProof>,
-    metadata: NoteMetadata,
-    recipient: Digest,
-    status: NoteStatus,
-}
-
-impl OutputNoteRecord {
-    pub fn new(
-        id: NoteId,
-        recipient: Digest,
-        assets: NoteAssets,
-        status: NoteStatus,
-        metadata: NoteMetadata,
-        inclusion_proof: Option<NoteInclusionProof>,
-        details: Option<NoteRecordDetails>,
-    ) -> OutputNoteRecord {
-        OutputNoteRecord {
-            id,
-            recipient,
-            assets,
-            status,
-            metadata,
-            inclusion_proof,
-            details,
-        }
-    }
-
-    pub fn id(&self) -> NoteId {
-        self.id
-    }
-
-    pub fn recipient(&self) -> Digest {
-        self.recipient
-    }
-
-    pub fn assets(&self) -> &NoteAssets {
-        &self.assets
-    }
-
-    pub fn status(&self) -> NoteStatus {
-        self.status
-    }
-
-    pub fn metadata(&self) -> &NoteMetadata {
-        &self.metadata
-    }
-
-    pub fn inclusion_proof(&self) -> Option<&NoteInclusionProof> {
-        self.inclusion_proof.as_ref()
-    }
-
-    pub fn details(&self) -> Option<&NoteRecordDetails> {
-        self.details.as_ref()
-    }
-}
-
-impl From<Note> for OutputNoteRecord {
-    fn from(note: Note) -> Self {
-        OutputNoteRecord {
-            id: note.id(),
-            recipient: note.recipient(),
-            assets: note.assets().clone(),
-            status: NoteStatus::Pending,
-            metadata: *note.metadata(),
-            inclusion_proof: None,
-            details: Some(NoteRecordDetails {
-                nullifier: note.nullifier().to_string(),
-                script: note.script().to_bytes(),
-                inputs: note.inputs().to_bytes(),
-                serial_num: note.serial_num(),
-            }),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
-pub struct NoteRecordDetails {
-    nullifier: String,
-    script: Vec<u8>,
-    inputs: Vec<u8>,
-    serial_num: Word,
-}
-
-impl NoteRecordDetails {
-    pub fn new(nullifier: String, script: Vec<u8>, inputs: Vec<u8>, serial_num: Word) -> Self {
-        Self {
-            nullifier,
-            script,
-            inputs,
-            serial_num,
-        }
-    }
-
-    pub fn script(&self) -> &Vec<u8> {
-        &self.script
-    }
-
-    pub fn inputs(&self) -> &Vec<u8> {
-        &self.inputs
-    }
-
-    pub fn serial_num(&self) -> Word {
-        self.serial_num
-    }
-}
-
-impl Serializable for NoteRecordDetails {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        let nullifier_bytes = self.nullifier.as_bytes();
-        target.write_usize(nullifier_bytes.len());
-        target.write_bytes(nullifier_bytes);
-
-        target.write_usize(self.script().len());
-        target.write_bytes(self.script());
-
-        target.write_usize(self.inputs().len());
-        target.write_bytes(self.inputs());
-
-        self.serial_num().write_into(target);
-    }
-}
-
-impl Deserializable for NoteRecordDetails {
-    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
-        let nullifier_len = usize::read_from(source)?;
-        let nullifier_bytes = source.read_vec(nullifier_len)?;
-        let nullifier =
-            String::from_utf8(nullifier_bytes).expect("Nullifier String bytes should be readable.");
-
-        let script_len = usize::read_from(source)?;
-        let script = source.read_vec(script_len)?;
-
-        let inputs_len = usize::read_from(source)?;
-        let inputs = source.read_vec(inputs_len)?;
-
-        let serial_num = Word::read_from(source)?;
-
-        Ok(NoteRecordDetails::new(
-            nullifier, script, inputs, serial_num,
-        ))
     }
 }
